:hammer: Refactor search input into separate component
